Text editor layout update
