+++ app/src/main/java/com/thesis/note/NavigationDrawer.kt
@@ -5,25 +5,25 @@ import android.content.Intent
 import android.view.MenuItem
 import androidx.core.view.GravityCompat
 import androidx.drawerlayout.widget.DrawerLayout
+import com.google.android.material.navigation.NavigationView
 import com.thesis.note.activity.*
 import com.thesis.note.test.DebugActivity
 
-//TODO documentation
-class NavigationDrawer(drawer_layout: DrawerLayout) {
-    val drawer_layout = drawer_layout
-
+/**
+ * [NavigationView] drawer logic
+ *
+ */
+class NavigationDrawer(val drawerLayout: DrawerLayout) {
+    /** On navigation item selected */
     fun onNavigationItemSelected(menuItem: MenuItem,context: Context): Boolean {
-
         when (menuItem.itemId) {
             R.id.start ->{
                 val intent = Intent(context, MainActivity::class.java)
                 intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
                 context.startActivity(intent)
             }
-
             R.id.drawer_list -> {
-                val listActivityIntent = Intent(context, ListActivity::class.java)
-
+                val listActivityIntent = Intent(context, MainActivity::class.java)
                 context.startActivity(listActivityIntent)
             }
             R.id.add_note ->{
@@ -38,19 +38,10 @@ class NavigationDrawer(drawer_layout: DrawerLayout) {
             R.id.drawer_debug ->
             {
                 val debugActivityIntent = Intent(context, DebugActivity::class.java)
-                context.startActivity(debugActivityIntent);
+                context.startActivity(debugActivityIntent)
             }
         }
-
-        drawer_layout.closeDrawer(GravityCompat.START)
-
+        drawerLayout.closeDrawer(GravityCompat.START)
         return true
     }
-
-
-
-
-
-
-
 }
