+++ app/src/main/java/com/thesis/note/activity/TextEditorActivity.kt
@@ -1,162 +1,262 @@
 package com.thesis.note.activity
 
 import android.content.Intent
+import android.graphics.Typeface
 import android.os.Bundle
 import android.text.Editable
-import android.view.MenuItem
-import android.view.View
-
 import android.widget.Toast
-import androidx.appcompat.app.ActionBarDrawerToggle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.GravityCompat
-import androidx.drawerlayout.widget.DrawerLayout
-import com.thesis.note.NavigationDrawer
-
-
-import com.thesis.note.database.AppDatabase
-import com.thesis.note.database.NoteType
-
-import com.google.android.material.navigation.NavigationView
-
-
-
+import androidx.core.content.res.ResourcesCompat
+import com.thesis.note.DrawerActivity
+import com.thesis.note.R
+import com.thesis.note.database.*
+import com.thesis.note.database.entity.Data
+import com.thesis.note.database.entity.Note
+import com.thesis.note.databinding.ActivityTextEditorLayoutBinding
+import com.thesis.note.fragment.ColorPickerFragment
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 
-import com.thesis.note.R
-import com.thesis.note.database.entity.Data
-import com.thesis.note.database.entity.Note
-import com.thesis.note.databinding.ActivityTextEditorBinding
+/**
+ * Activity for text editing.
+ *
+ * When creating [Intent] of this activity, you should put extended data with
+ * putExtra("noteID", yourNoteID) and putExtra("dataID", yourDataID)
+ *
+ * TODO on back pressed dialog (discard changes)
+ */
+class TextEditorActivity : DrawerActivity() {
+    /** This activity */
+    private val thisActivity = this
 
-@Deprecated("Use TextEditorNewActivity")
-class TextEditorActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
-    lateinit var drawer_layout: DrawerLayout
-    lateinit var navigationDrawer : NavigationDrawer
-    private lateinit var binding: ActivityTextEditorBinding
+    /** View binding */
+    lateinit var binding: ActivityTextEditorLayoutBinding
 
-    var dataExistInDB:Boolean = false
-    var noteExistInDB:Boolean = false
-    var dataID:Int = -1
-    var noteID:Int = -1
+    /** Database */
+    lateinit var db: AppDatabase
 
-    val TextEditorActivityContext = this
+    /** Edited [Data] id */
+    private var dataID:Int = -1
+    /** Edited [Data] */
+    private lateinit var editedData: Data
+    /** Edited [Note]  id */
+    private var noteID:Int = -1
+    /** Edited [Note] */
+    private lateinit var editedNote: Note
 
-    lateinit var editedData: Data
+    /** */
+    private var italic = false
+    /** */
+    private var bold = false
+    /** */
+    private var fontSize = 16
+    /** */
+    private var fontColor = NoteColor.Black
 
-    val db = AppDatabase.invoke(this)
-
-    lateinit var noteViewerActivityIntent : Intent
-
+    /** On create callback */
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityTextEditorBinding.inflate(layoutInflater) //LAYOUT BINDING CLASS
-        setContentView(binding.root)
-        drawer_layout = binding.activityTextEditorLayout
-        navigationDrawer = NavigationDrawer(drawer_layout)
-        binding.navigationView.setNavigationItemSelectedListener(this)
-
-        val drawerToggle= ActionBarDrawerToggle(this,drawer_layout,binding.toolbar,R.string.abdt,R.string.abdt)
-        drawer_layout.addDrawerListener(drawerToggle)
-        drawerToggle.isDrawerIndicatorEnabled = true
-        drawerToggle.syncState()
-        //------------------------------------------------------------------------------------------
-        noteViewerActivityIntent = Intent(this, NoteViewerActivity::class.java)
-        //Check if new data or edit of existing
-        val parameters = intent.extras
-
-        if(parameters != null) {
-            dataID = parameters.getInt("dataID")
-            noteID = parameters.getInt("noteID")
-
-            if(dataID <= 0){
-                dataExistInDB = false
-            }
-            else{
-                //load data
-                GlobalScope.launch {
-                    editedData = db.dataDao().getDataById(dataID)
-                    dataExistInDB = true
-                    //show data in textField
-                    TextEditorActivityContext.runOnUiThread(
-                        fun(){
-                            binding.textField.text = Editable.Factory.getInstance().newEditable(editedData.Content)
-                        })}}
-            //check if new note
-            noteExistInDB = noteID == -1
-                }
-        else{
-            dataExistInDB = false
-            noteExistInDB = false
+        binding = ActivityTextEditorLayoutBinding.inflate(layoutInflater)
+        setDrawerLayout(binding.root,binding.toolbar,binding.navigationView)
+        db = AppDatabase.invoke(this)
+        loadParameters()
+        GlobalScope.launch {
+            loadData()
+            setData()
         }
 
-        //SAVE button
-        binding.saveButton.setOnClickListener(object : View.OnClickListener{
-            override fun onClick(v: View?) {
-                if(dataExistInDB){
-                    //Update data
-                    editedData.Content = binding.textField.text.toString()
-                    GlobalScope.launch {
-                        db.dataDao().update(editedData)
-                    }
+        //Save button listener
+        binding.saveButton.setOnClickListener {
+            if (dataID != -1) {
+                //Update data
+                GlobalScope.launch {
+                    db.dataDao().update(editedData.apply{
+                        Content = binding.editedText.text.toString()
+                        Info = getInfo()
+                        Color = fontColor
+                        Size = fontSize
+                    })
                 }
-                else{
-                    if(noteExistInDB){
+            } else {
+                if (noteID != -1) {
                     //Add new data to database
                     GlobalScope.launch {
-                        val newDataID = db.dataDao().insertAll(Data(0,noteID,NoteType.Text,binding.textField.text.toString(),null,null,null))
+                        db.dataDao().insertAll(Data(0, noteID, NoteType.Text, binding.editedText.text.toString(), getInfo(),fontSize,fontColor))
+                    }
+                } else {
+                    GlobalScope.launch {
+                        //add new note
+                        val idNewNote =
+                            db.noteDao().insertAll(Note(0, "", null, null, false, null, null, null, NoteColor.White))
+                        noteID = idNewNote[0].toInt()
+                        //add new data
+                        val newDataID = db.dataDao().insertAll(Data(0, noteID, NoteType.Text, binding.editedText.text.toString(), getInfo(),fontSize,fontColor))
                         dataID = newDataID[0].toInt()
-                        dataExistInDB = true
-                    }}
-                    else{
-                        GlobalScope.launch {
-                            //add new note
-                            var idNewNote = db.noteDao().insertAll(Note(0,"",null,null,false,null,null,null,null))
-                            noteID = idNewNote[0].toInt()
-                            //add new data
-                            val newDataID = db.dataDao().insertAll(Data(0,noteID,NoteType.Text,binding.textField.text.toString(),null,null,null))
-                            dataID = newDataID[0].toInt()
-                            val note = db.noteDao().getNoteById(noteID)
-                            note.MainData = dataID
-                            db.noteDao().update(note)
-                            //open new note
-                            noteViewerActivityIntent.putExtra("noteID",noteID)
-                            TextEditorActivityContext.startActivity(noteViewerActivityIntent)
+                        editedNote = db.noteDao().getNoteById(noteID)
+                        db.noteDao().update(editedNote.apply { MainData = dataID })
+                        //open new note
+                        runOnUiThread {
+                            thisActivity.startActivity(Intent(thisActivity, NoteViewerActivity::class.java).apply{putExtra("noteID", noteID)})
                         }
-
-
-
-
                     }
                 }
-                Toast.makeText(applicationContext,"ZAPISANO", Toast.LENGTH_SHORT).show()
-                finish()
-            }})
-
-        //REMOVE button
-        binding.deleteButton.setOnClickListener(object : View.OnClickListener{
-            override fun onClick(v: View?) {
-                if(dataExistInDB)
-                GlobalScope.launch {
-                    val db = AppDatabase(applicationContext)
-                    db.dataDao().delete(editedData)
-                }
-                Toast.makeText(applicationContext,"USUNIĘTO", Toast.LENGTH_SHORT).show()
-                finish()
             }
-        })
+            Toast.makeText(applicationContext, R.string.activity_text_editor_save_OK, Toast.LENGTH_SHORT).show()
+            finish()
+        }
+
+        //TODO Delete button listener
+        binding.deleteButton.setOnClickListener {
+            Toast.makeText(applicationContext, R.string.not_implemented, Toast.LENGTH_SHORT).show()
+        }
+
+        //TODO Share button listener
+        binding.shareButton.setOnClickListener {
+            Toast.makeText(applicationContext, R.string.not_implemented, Toast.LENGTH_SHORT).show()
+        }
+
+        //Text color button listener
+        binding.textColorButton.setOnClickListener {
+            ColorPickerFragment(ColorPalette.TEXT_COLOR_PALETTE).show(supportFragmentManager, "tag")
+        }
+
+        //ColorPickerFragment result listener
+        supportFragmentManager.setFragmentResultListener("color",this) { _, bundle ->
+            val result = bundle.getInt("colorID")
+            fontColor = NoteColorConverter().intToEnum(result)!!
+            binding.editedText.setTextColor(resources.getColor(NoteColorConverter.enumToColor(fontColor),null))
+        }
+
+        //TODO Text size button listener
+        binding.textSizeButton.setOnClickListener {
+            Toast.makeText(applicationContext, R.string.not_implemented, Toast.LENGTH_SHORT).show()
+        }
+
+        //Italic button listener
+        binding.italicTextButton.setOnClickListener{
+            if(italic){
+                setItalicText(false)
+            }
+            else{
+                setItalicText(true)
+            }
+        }
+
+        //Bold button listener
+        binding.boldTextButton.setOnClickListener {
+            if(bold){
+                setBoldText(false)
+            }
+            else{
+                setBoldText(true)
+            }
+        }
+
     }
 
-    override fun onNavigationItemSelected(menuItem: MenuItem): Boolean {
-        finish()
-        return navigationDrawer.onNavigationItemSelected(menuItem,this)
-    }
-
-    override fun onBackPressed() {
-        if (drawer_layout.isDrawerOpen(GravityCompat.START)) {
-            drawer_layout.closeDrawer(GravityCompat.START)
+    /** Load parameters passed from another activity */
+    private fun loadParameters() {
+        val parameters = intent.extras
+        if (parameters == null) {
+            Toast.makeText(
+                applicationContext,
+                R.string.activity_text_editor_error_parameters,
+                Toast.LENGTH_SHORT
+            ).show()
         } else {
-            super.onBackPressed()
+            dataID = parameters.getInt("dataID")
+            noteID = parameters.getInt("noteID")
         }
     }
+
+    /** Load [Note] with id [noteID] and [Data] with id [dataID]  */
+    private fun loadData(){
+        if(dataID != -1){
+            editedData = db.dataDao().getDataById(dataID)
+            fontSize = editedData.Size!!
+            fontColor = editedData.Color!!
+            when(editedData.Info){
+                "B" -> bold = true
+                "I" -> italic = true
+                "BI" -> {
+                    bold = true
+                    italic = true}
+            }
+        }
+        if(noteID != -1){
+            editedNote = db.noteDao().getNoteById(noteID)
+        }
+    }
+
+    /** Set loaded [Data] into layout */
+    private fun setData(){
+        //show data in textField
+        setText(editedData.Content)
+        //load graphic options
+        setItalicText(italic)
+        setBoldText(bold)
+        binding.editedText.textSize = fontSize.toFloat()
+        binding.editedText.setTextColor(resources.getColor(NoteColorConverter.enumToColor(fontColor),null))
+        //set background
+        binding.editedText.background = ResourcesCompat.getDrawable(resources, NoteColorConverter.enumToColor(editedNote.Color),null)
+    }
+
+    /** Get info about italic and bold */
+    private fun getInfo(): String? {
+        if(bold && italic)
+            return "BI"
+        if(bold)
+            return "B"
+        if(italic)
+            return "I"
+        return null
+    }
+
+    /** Set italic text when [value] is true, disable italic text when [value] is false */
+    private fun setItalicText(value: Boolean) {
+        italic = value
+        if(value){
+            if(bold){
+                setBoldItalicText()
+            }else{
+                binding.editedText.typeface = Typeface.create(Typeface.DEFAULT, Typeface.ITALIC)
+            }
+        }
+        else{
+            if(bold){
+                setBoldText(true)
+            }else{
+                binding.editedText.typeface = Typeface.create(Typeface.DEFAULT, Typeface.NORMAL)
+            }
+        }
+    }
+
+    /** Set bold text when [value] is true, disable bold text when [value] is false */
+    private fun setBoldText(value: Boolean) {
+        bold = value
+        if(value){
+            if(italic){
+                setBoldItalicText()
+            }else{
+                binding.editedText.typeface = Typeface.create(Typeface.DEFAULT, Typeface.BOLD)
+            }
+        }
+        else{
+            if(italic){
+                setItalicText(true)
+            }else{
+                binding.editedText.typeface = Typeface.create(Typeface.DEFAULT, Typeface.NORMAL)
+            }
+        }
+    }
+
+    /** Set text to bold and italic */
+    private fun setBoldItalicText() {
+        binding.editedText.typeface = Typeface.create(Typeface.DEFAULT, Typeface.BOLD_ITALIC)
+    }
+
+    /** Set content into text edit */
+    private fun setText(content: String?){
+        binding.editedText.text = Editable.Factory.getInstance().newEditable(content)
+    }
+
 }
