+++ app/src/main/res/layout/drawer_header.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!-- TODO -->
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="192dp"
@@ -7,9 +6,12 @@
     android:gravity="bottom"
     android:orientation="vertical"
     android:padding="16dp"
-    android:theme="@style/ThemeOverlay.AppCompat.Dark"><TextView
-android:layout_width="match_parent"
+    android:theme="@style/ThemeOverlay.AppCompat.Dark">
+
+    <TextView
+    android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:textSize="30sp"
     android:text="@string/app_name"
     android:textAppearance="@style/TextAppearance.AppCompat.Body1" />
     </LinearLayout>
