+++ app/src/main/res/values/strings.xml
@@ -2,27 +2,8 @@
     <!-- TODO -->
 
     <!-- Old -->
-    <string name="app_name">Notes</string>
+
     <string name="abdt">ActionBarDrawerToggle</string>
-    <string-array name="groups_array">
-        <item>Bez grupy</item>
-        <item>Grupa 1</item>
-        <item>Grupa 2</item>
-        <item>Grupa 3</item>
-    </string-array>
-    <string-array name="font_name">
-        <item>Nazwa czcionki</item>
-    </string-array>
-    <string-array name="font_size">
-        <item>Rozmiar czcionki</item>
-    </string-array>
-    <string-array name="font_color">
-        <item>Kolor czcionki</item>
-    </string-array>
-    <string-array name="background_color">
-        <item>Kolor tła</item>
-    </string-array>
-    <string name="error_empty">ERROR_EMPTY</string>
     <string name="note_type_text">Text</string>
     <string name="note_type_list">List</string>
     <string name="note_type_video">Video</string>
@@ -50,14 +31,23 @@
     <string name="add_note_image">Add image note</string>
     <string name="add_note_title">Add new note</string>
     <string name="no_permissions">No permissions</string>
-    <string name="action_bar_drawer_toggle_open">Open drawer</string>
-    <string name="action_bar_drawer_toggle_close">Close drawer</string>
-    <string name="groups_without_group">R.string.groups_without_group CHANGE</string>
+
 
     <!-- Temp -->
     <string name="not_implemented">Not implemented</string>
 
     <!-- New -->
+    <string name="app_name">Notes</string>
+
+    <string name="drawer_menu_all">All notes</string>
+    <string name="drawer_menu_favorites">Favorites</string>
+    <string name="drawer_menu_add">Add note</string>
+    <string name="drawer_menu_settings">Settings</string>
+    <string name="drawer_menu_debug" translatable="false">Debug</string>
+
+    <string name="action_bar_drawer_toggle_open">Open drawer</string>
+    <string name="action_bar_drawer_toggle_close">Close drawer</string>
+
     <string name="x_template_empty_activity_template">TEMPLATE</string>
 
     <string name="activity_main_search_button_description">Search notes</string>
@@ -81,5 +71,14 @@
 
     <string name="recycler_view_tag_list_error">ERROR: empty</string>
 
+    <string name="activity_text_editor_save_button_description">Save</string>
+    <string name="activity_text_editor_delete_button_description">Delete</string>
+    <string name="activity_text_editor_share_button_description">Share</string>
+    <string name="activity_text_editor_text_color_button_description">Color</string>
+    <string name="activity_text_editor_text_size_button_description">Size</string>
+    <string name="activity_text_editor_italic_button_description">Italic</string>
+    <string name="activity_text_editor_bold_button_description">Bold</string>
+    <string name="activity_text_editor_error_parameters" translatable="false">ERROR: bad note ID</string>
+    <string name="activity_text_editor_save_OK">Data saved</string>
 
 </resources>
